+++ lib/features/authorization/presentation/pages/number_sign_up.dart
@@ -10,7 +10,7 @@ class NumberSignUp extends StatefulWidget {
 class _NumberSignUpState extends State<NumberSignUp> {
   TextEditingController _textEditingController;
   FocusNode focusNode;
-  bool _disable = true;
+
   String _countryCode = 'NG';
   void showCountryCodeDialoge() => showDialog(
       context: context,
@@ -25,6 +25,9 @@ class _NumberSignUpState extends State<NumberSignUp> {
   @override
   void initState() {
     _textEditingController = TextEditingController();
+    _textEditingController.addListener(() {
+      setState(() {});
+    });
     // focusNode = FocusNode();
     // focusNode.requestFocus();
     super.initState();
@@ -41,119 +44,114 @@ class _NumberSignUpState extends State<NumberSignUp> {
   @override
   Widget build(BuildContext context) {
     var _size = MediaQuery.of(context).size;
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.center,
-      children: [
-        Container(
-            width: _size.width,
-            height: 50.0,
-            margin: EdgeInsets.symmetric(horizontal: 20.0),
-            padding: EdgeInsets.only(left: 20.0),
-            decoration: BoxDecoration(
-              borderRadius: BorderRadius.circular(6.0),
-              border: Border.all(
-                color: Colors.grey,
-              ),
-              color: Colors.grey.shade300,
-            ),
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.start,
-              children: [
-                // Text(
-                //   _countryCode,
-                //   style: TextStyle(fontWeight: FontWeight.bold),
-                // ),
-                CountryCodePicker(
-                  padding: EdgeInsets.all(0.0),
-
-                  onChanged: (CountryCode countryCode) {
-                    setState(() {
-                      _countryCode = countryCode.code;
-                    });
-                  },
-
-                  // showOnlyCountryWhenClosed: false,
-                  // showFlagDialog: false,
-                  // showFlagMain: false,
-                  // showCountryOnly: true,
-
-                  //showDropDownButton: true,
-                  showFlag: true,
-                  //dd
-                  builder: (CountryCode countrycode) {
-                    return Row(
-                      mainAxisAlignment: MainAxisAlignment.spaceAround,
-                      children: [
-                        Text(countrycode.code,
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                            )),
-                        SizedBox(
-                          width: 5.0,
-                        ),
-                        Text(countrycode.dialCode,
-                            style: TextStyle(
-                              fontWeight: FontWeight.bold,
-                            ))
-                      ],
-                    );
-                  },
+    return SingleChildScrollView(
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.center,
+        children: [
+          Container(
+              width: _size.width,
+              height: 50.0,
+              padding: EdgeInsets.only(left: 20.0),
+              decoration: BoxDecoration(
+                borderRadius: BorderRadius.circular(6.0),
+                border: Border.all(
+                  color: Colors.grey,
                 ),
-                Container(
-                    margin: EdgeInsets.symmetric(
-                      vertical: 10.0,
-                    ),
-                    child: VerticalDivider()),
-                Expanded(
-                  child: TextField(
-                    // controller: _textEditingController,
-                    onChanged: (value) {
+                color: Colors.grey.shade300,
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.start,
+                children: [
+                  // Text(
+                  //   _countryCode,
+                  //   style: TextStyle(fontWeight: FontWeight.bold),
+                  // ),
+                  CountryCodePicker(
+                    padding: EdgeInsets.all(0.0),
+
+                    onChanged: (CountryCode countryCode) {
                       setState(() {
-                        if (value.isNotEmpty) {
-                          _disable = false;
-                        } else {
-                          _disable = true;
-                        }
+                        _countryCode = countryCode.code;
                       });
                     },
-                    // autofocus: true,
-                    // showCursor: true,
 
-                    //keyboardType: TextInputType.number,
-                    decoration: InputDecoration(
-                      // suffixIcon: _disable
-                      //     ? Container()
-                      //     : IconButton(
-                      //         icon: Icon(
-                      //           Icons.close,
-                      //           color: Colors.grey,
-                      //         ),
-                      //         onPressed: () {
-                      //           _textEditingController.clear();
-                      //         },
-                      //       ),
-                      hintText: 'Phone number',
-                      border: InputBorder.none,
+                    // showOnlyCountryWhenClosed: false,
+                    // showFlagDialog: false,
+                    // showFlagMain: false,
+                    // showCountryOnly: true,
+
+                    //showDropDownButton: true,
+                    showFlag: true,
+                    //dd
+                    builder: (CountryCode countrycode) {
+                      return Row(
+                        mainAxisAlignment: MainAxisAlignment.spaceAround,
+                        children: [
+                          Text(countrycode.code,
+                              style: TextStyle(
+                                fontWeight: FontWeight.bold,
+                              )),
+                          SizedBox(
+                            width: 5.0,
+                          ),
+                          Text(countrycode.dialCode,
+                              style: TextStyle(
+                                fontWeight: FontWeight.bold,
+                              ))
+                        ],
+                      );
+                    },
+                  ),
+                  Container(
+                      margin: EdgeInsets.symmetric(
+                        vertical: 10.0,
+                      ),
+                      child: VerticalDivider()),
+                  Expanded(
+                    child: TextField(
+                      controller: _textEditingController,
+                      onChanged: (value) {
+                        setState(() {});
+                      },
+                      // autofocus: true,
+
+                      keyboardType: TextInputType.number,
+                      decoration: InputDecoration(
+                        hintText: 'Phone number',
+                        border: InputBorder.none,
+                      ),
                     ),
                   ),
-                ),
-              ],
-            )),
-        SizedBox(
-          height: 10.0,
-        ),
-        Text(
-          'You may receive SMS updates from instagram and can opt out any time',
-          textAlign: TextAlign.center,
-        ),
-        SizedBox(
-          height: 10.0,
-        ),
-        LoginProceedButton(
-          child: Text('Next'),
-          disableButton: _disable,
-        )
-      ],
+                  _textEditingController.text.isNotEmpty
+                      ? IconButton(
+                          splashColor: Colors.transparent,
+                          icon: Icon(
+                            Icons.close,
+                            color: Colors.grey,
+                          ),
+                          onPressed: () {
+                            _textEditingController.clear();
+                          },
+                        )
+                      : Container(),
+                ],
+              )),
+          SizedBox(
+            height: 15.0,
+          ),
+          Text(
+            'You may receive SMS updates from instagram and can opt out any time',
+            textAlign: TextAlign.center,
+          ),
+          SizedBox(
+            height: 15.0,
+          ),
+          LoginProceedButton(
+            child: Text('Next'),
+            disableButton: _textEditingController.text.isEmpty,
+          )
+        ],
+      ),
     );
   }
 }
