+++ lib/main.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:instagram_clone/features/authorization/presentation/pages/welcome_page.dart';
+import 'package:instagram_clone/features/authorization/presentation/pages/login_page.dart';
 import 'package:instagram_clone/features/home/presentation/pages/home.dart';
 import 'package:instagram_clone/features/profile/presentation/pages/profile_page.dart';
+import 'package:instagram_clone/features/authorization/presentation/pages/sign_up_page.dart';
 
 void main() {
   runApp(MyApp());
@@ -48,7 +49,7 @@ class MyApp extends StatelessWidget {
         ),
         bottomAppBarColor: Color(0xFF000000),
       ),
-      home: WelcomePage(),
+      home: LoginPage(),
     );
   }
 }
