+++ pubspec.yaml
@@ -26,6 +26,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   language_picker: ^0.4.1
   country_code_picker: ^2.0.2
+  font_awesome_flutter: ^9.1.0
 
 dev_dependencies:
   flutter_test:
